+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Features
 
+- Adds support for setting power schedules for seed scheduling in
+  [#73](https://github.com/TNO-S3/WuppieFuzz/pull/73)
 - Adds custom executor enabling proper timeout kill and removing unsafe code in
   [#49](https://github.com/TNO-S3/WuppieFuzz/pull/49)